requirements.txt
altgraph==0.17.3
click==8.1.3
Flask==2.2.3
imagezmq==1.1.1
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
opencv-python==4.7.0.72
pycryptodomex==3.17
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
pyserial==3.5
pyzmq==25.0.0
simplejpeg==1.6.5
Werkzeug==2.2.3
